requirements.txt
aniso8601==1.1.0
asn1crypto==0.24.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.6
cryptography==2.9.2
EditorConfig==0.12.2
enum34==1.1.6
Fabric==1.13.1
Flask==0.11.1
Flask-Cache==0.13.1
Flask-Cors==3.0.2
Flask-HTTPAuth==3.2.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
idna==2.5
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.8
jsbeautifier==1.6.14
jsmin==2.2.2
jsonschema==2.5.1
MarkupSafe==0.23
mock==2.0.0
numpy==1.12.1
pandas==0.24.2
paramiko==2.2.1
pbr==5.1.2
pyasn1==0.4.6
pycparser==2.20
pycrypto==2.6.1
PyMySQL==0.7.9
PyNaCl==1.3.0
python-dateutil==2.5.3
python-ldap==3.2.0
pytz==2016.6.1
redis==2.10.6
requests==2.18.1
simplejson==3.9.0
six==1.10.0
SQLAlchemy==1.1.13
Unidecode==0.4.21
urllib3==1.21.1
Werkzeug==0.11.11
WTForms==2.1
xlsxwriter==1.2.0